split provider/common.go into platform.go and platform_validations.go
